+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Grayscale effect.
 
 ### Changed
+- Keyframes are now entered as `new vd.KeyFrame([time1, val1, interp],
+  [time2, val2])`
 - Rename `cctx` ('canvas context') to `vctx` ('visual context').
 
 ### Removed